python/plotly-intro/03_Simple_BigQuery/requirements.txt
dash==1.7.0
Flask==1.1.1
gunicorn==20.0.4
google-cloud-bigquery==1.22.0
pandas==0.25.*